+++ extras/arp_spoofer/Makefile
@@ -1,0 +1,85 @@
+# Get the currently used golang install path (in GOPATH/bin, unless GOBIN is set)
+ifeq (,$(shell go env GOBIN))
+GOBIN=$(shell go env GOPATH)/bin
+else
+GOBIN=$(shell go env GOBIN)
+endif
+
+ifneq ("$(wildcard ./version)","")
+	ARPSPOOFER_VERSION ?= $(shell cat ./version)
+else ifeq ($(shell git describe --exact-match --tags 2> /dev/null),)
+	ARPSPOOFER_VERSION ?= $(shell git rev-parse --short HEAD)
+else
+	ARPSPOOFER_VERSION ?= $(shell (git describe --exact-match --tags 2> /dev/null | cut -dv -f2))
+endif
+
+# Setting SHELL to bash allows bash commands to be executed by recipes.
+# This is a requirement for 'setup-envtest.sh' in the test target.
+# Options are set to exit when a recipe line exits non-zero or a piped command fails.
+SHELL = /usr/bin/env bash -o pipefail
+.SHELLFLAGS = -ec
+
+all: build
+
+##@ General
+
+# The help target prints out all targets with their descriptions organized
+# beneath their categories. The categories are represented by '##@' and the
+# target descriptions by '##'. The awk commands is responsible for reading the
+# entire set of makefiles included in this invocation, looking for lines of the
+# file as xyz: ## something, and then pretty-format the target and help. Then,
+# if there's a line with ##@ something, that gets pretty-printed as a category.
+# More info on the usage of ANSI control characters for terminal formatting:
+# https://en.wikipedia.org/wiki/ANSI_escape_code#SGR_parameters
+# More info on the awk command:
+# http://linuxcommand.org/lc3_adv_awk.php
+
+help: ## Display this help.
+	@awk 'BEGIN {FS = ":.*##"; printf "\nUsage:\n  make \033[36m<target>\033[0m\n"} /^[a-zA-Z_0-9-]+:.*?##/ { printf "  \033[36m%-15s\033[0m %s\n", $$1, $$2 } /^##@/ { printf "\n\033[1m%s\033[0m\n", substr($$0, 5) } ' $(MAKEFILE_LIST)
+
+##@ Development
+
+.PHONY: fmt
+fmt: ## Run go fmt against code.
+	go fmt ./...
+
+.PHONY: lint
+lint:
+	golangci-lint run ./...
+
+.PHONY: vet
+vet:
+	go vet ./...
+
+.PHONY: addlicense
+addlicense: ## Add license headers to all go files.
+	find . -name '*.go' -exec go run github.com/google/addlicense -c 'IronCore authors' {} +
+
+.PHONY: checklicense
+checklicense: ## Check that every file has a license header present.
+	find . -name '*.go' -exec go run github.com/google/addlicense  -check -c 'IronCore authors' {} +
+
+.PHONY: check
+check: addlicense lint test # Generate manifests, code, lint, add licenses, test
+
+.PHONY: test
+test: fmt vet ## Run tests.
+	go test ./... -coverprofile cover.out
+
+##@ Build
+
+.PHONY: build
+build: fmt vet ## Build binary.
+	go build -ldflags "-X 'main.version=$(ARPSPOOFER_VERSION)'" -o bin/arp_spoofer main.go
+
+.PHONY: install
+install:
+	go install .
+
+.PHONY: run
+run: fmt lint ## Run a controller from your host.
+	go run ./main.go
+
+.PHONY: release
+release: fmt lint
+	goreleaser release --snapshot --rm-dist
