+++ extras/arp_spoofer/go.mod
@@ -1,0 +1,16 @@
+module github.com/ironcore-dev/metalbond/extra/arp_spoofer
+
+go 1.22.3
+
+require (
+	github.com/alecthomas/kong v1.9.0
+	github.com/google/gopacket v1.1.19
+	github.com/sirupsen/logrus v1.9.3
+)
+
+require (
+	github.com/bmatcuk/doublestar/v4 v4.0.2 // indirect
+	github.com/google/addlicense v1.1.1 // indirect
+	golang.org/x/sync v0.0.0-20190911185100-cd5d95a43a6e // indirect
+	golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
+)
